+++ configs/swin_transformer_v2/swinv2-base_16xb64_window16_in1k.py
@@ -1,8 +1,8 @@
 _base_ = [
-    '../_base_/models/swin_transformer_v2/small_256.py',
+    '../_base_/models/swin_transformer_v2/base_256.py',
     '../_base_/datasets/imagenet_bs64_swin_256.py',
     '../_base_/schedules/imagenet_bs1024_adamw_swin.py',
     '../_base_/default_runtime.py'
 ]
 
-model = dict(backbone=dict(window_size=[16, 16, 16, 8]), )
+model = dict(backbone=dict(window_size=[16, 16, 16, 8]))
