+++ configs/swin_transformer_v2/swinv2-large_16xb64_window12to16_192to256_in22k.py
@@ -1,5 +1,6 @@
+# Only for evaluation
 _base_ = [
-    '../_base_/models/swin_transformer_v2/base_256.py',
+    '../_base_/models/swin_transformer_v2/large_256.py',
     '../_base_/datasets/imagenet_bs64_swin_256.py',
     '../_base_/schedules/imagenet_bs1024_adamw_swin.py',
     '../_base_/default_runtime.py'
@@ -7,5 +8,6 @@ _base_ = [
 
 model = dict(
     type='ImageClassifier',
-    backbone=dict(window_size=[16, 16, 16, 8]),
-    drop_path_rate=0.2)
+    backbone=dict(
+        window_size=[16, 16, 16, 8], pretrained_window_sizes=[12, 12, 12, 6]),
+)
