+++ configs/swin_transformer_v2/swinv2-large_16xb64_window12to24_192to384_in22k.py
@@ -8,5 +8,8 @@ _base_ = [
 
 model = dict(
     type='ImageClassifier',
-    backbone=dict(img_size=384, window_size=[24, 24, 24, 12]),
+    backbone=dict(
+        img_size=384,
+        window_size=[24, 24, 24, 12],
+        pretrained_window_sizes=[12, 12, 12, 6]),
 )
