+++ mmcls/models/utils/embed.py
@@ -1,11 +1,13 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 import warnings
+from typing import Sequence
 
 import numpy as np
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from mmcv.cnn import build_conv_layer, build_norm_layer
+from mmcv.cnn.bricks.transformer import AdaptivePadding
 from mmcv.runner.base_module import BaseModule
 
 from .helpers import to_2tuple
@@ -272,93 +274,147 @@ class HybridEmbed(BaseModule):
 
 
 class PatchMerging(BaseModule):
-    """Merge patch feature map.
+    """Merge patch feature map. Modified from mmcv, which uses pre-norm layer
+    whereas Swin V2 uses post-norm here. Therefore, add extra parameter to
+    decide whether use post-norm or not.
 
-    This layer use nn.Unfold to group feature map by kernel_size, and use norm
-    and linear layer to embed grouped feature map.
+    This layer groups feature map by kernel_size, and applies norm and linear
+    layers to the grouped feature map ((used in Swin Transformer)).
+    Our implementation uses `nn.Unfold` to
+    merge patches, which is about 25% faster than the original
+    implementation. However, we need to modify pretrained
+    models for compatibility.
 
     Args:
-        input_resolution (tuple): The size of input patch resolution.
         in_channels (int): The num of input channels.
-        expansion_ratio (Number): Expansion ratio of output channels. The num
-            of output channels is equal to int(expansion_ratio * in_channels).
+            to gets fully covered by filter and stride you specified.
+        out_channels (int): The num of output channels.
         kernel_size (int | tuple, optional): the kernel size in the unfold
             layer. Defaults to 2.
         stride (int | tuple, optional): the stride of the sliding blocks in the
-            unfold layer. Defaults to be equal with kernel_size.
-        padding (int | tuple, optional): zero padding width in the unfold
-            layer. Defaults to 0.
+            unfold layer. Defaults to None. (Would be set as `kernel_size`)
+        padding (int | tuple | string ): The padding length of
+            embedding conv. When it is a string, it means the mode
+            of adaptive padding, support "same" and "corner" now.
+            Defaults to "corner".
         dilation (int | tuple, optional): dilation parameter in the unfold
-            layer. Defaults to 1.
+            layer. Default: 1.
         bias (bool, optional): Whether to add bias in linear layer or not.
             Defaults to False.
         norm_cfg (dict, optional): Config dict for normalization layer.
             Defaults to dict(type='LN').
+        is_post_norm (bool): Whether to use post normalization here.
+            Defaults to False.
         init_cfg (dict, optional): The extra config for initialization.
             Defaults to None.
     """
 
     def __init__(self,
-                 input_resolution,
                  in_channels,
-                 expansion_ratio,
+                 out_channels,
                  kernel_size=2,
                  stride=None,
-                 padding=0,
+                 padding='corner',
                  dilation=1,
                  bias=False,
                  norm_cfg=dict(type='LN'),
+                 is_post_norm=False,
                  init_cfg=None):
-        super().__init__(init_cfg)
-        warnings.warn('The `PatchMerging` in mmcls will be deprecated. '
-                      'Please use `mmcv.cnn.bricks.transformer.PatchMerging`. '
-                      "It's more general and supports dynamic input shape")
-
-        H, W = input_resolution
-        self.input_resolution = input_resolution
+        super().__init__(init_cfg=init_cfg)
         self.in_channels = in_channels
-        self.out_channels = int(expansion_ratio * in_channels)
+        self.out_channels = out_channels
+        self.is_post_norm = is_post_norm
 
-        if stride is None:
+        if stride:
+            stride = stride
+        else:
             stride = kernel_size
+
         kernel_size = to_2tuple(kernel_size)
         stride = to_2tuple(stride)
-        padding = to_2tuple(padding)
         dilation = to_2tuple(dilation)
-        self.sampler = nn.Unfold(kernel_size, dilation, padding, stride)
+
+        if isinstance(padding, str):
+            self.adaptive_padding = AdaptivePadding(
+                kernel_size=kernel_size,
+                stride=stride,
+                dilation=dilation,
+                padding=padding)
+            # disable the padding of unfold
+            padding = 0
+        else:
+            self.adaptive_padding = None
+
+        padding = to_2tuple(padding)
+        self.sampler = nn.Unfold(
+            kernel_size=kernel_size,
+            dilation=dilation,
+            padding=padding,
+            stride=stride)
 
         sample_dim = kernel_size[0] * kernel_size[1] * in_channels
 
+        self.reduction = nn.Linear(sample_dim, out_channels, bias=bias)
+
         if norm_cfg is not None:
-            self.norm = build_norm_layer(norm_cfg, sample_dim)[1]
+            # build pre or post norm layer based on different channels
+            if self.is_post_norm:
+                self.norm = build_norm_layer(norm_cfg, out_channels)[1]
+            else:
+                self.norm = build_norm_layer(norm_cfg, sample_dim)[1]
         else:
             self.norm = None
 
-        self.reduction = nn.Linear(sample_dim, self.out_channels, bias=bias)
-
-        # See https://pytorch.org/docs/stable/generated/torch.nn.Conv2d.html
-        H_out = (H + 2 * padding[0] - dilation[0] *
-                 (kernel_size[0] - 1) - 1) // stride[0] + 1
-        W_out = (W + 2 * padding[1] - dilation[1] *
-                 (kernel_size[1] - 1) - 1) // stride[1] + 1
-        self.output_resolution = (H_out, W_out)
-
-    def forward(self, x):
+    def forward(self, x, input_size):
         """
-        x: B, H*W, C
+        Args:
+            x (Tensor): Has shape (B, H*W, C_in).
+            input_size (tuple[int]): The spatial shape of x, arrange as (H, W).
+                Default: None.
+
+        Returns:
+            tuple: Contains merged results and its spatial shape.
+
+            - x (Tensor): Has shape (B, Merged_H * Merged_W, C_out)
+            - out_size (tuple[int]): Spatial shape of x, arrange as
+              (Merged_H, Merged_W).
         """
-        H, W = self.input_resolution
         B, L, C = x.shape
+        assert isinstance(input_size, Sequence), f'Expect ' \
+                                                 f'input_size is ' \
+                                                 f'`Sequence` ' \
+                                                 f'but get {input_size}'
+
+        H, W = input_size
         assert L == H * W, 'input feature has wrong size'
 
         x = x.view(B, H, W, C).permute([0, 3, 1, 2])  # B, C, H, W
 
+        if self.adaptive_padding:
+            x = self.adaptive_padding(x)
+            H, W = x.shape[-2:]
+
         # Use nn.Unfold to merge patch. About 25% faster than original method,
         # but need to modify pretrained model for compatibility
-        x = self.sampler(x)  # B, 4*C, H/2*W/2
+        # if kernel_size=2 and stride=2, x should has shape (B, 4*C, H/2*W/2)
+        x = self.sampler(x)
+
+        out_h = (H + 2 * self.sampler.padding[0] - self.sampler.dilation[0] *
+                 (self.sampler.kernel_size[0] - 1) -
+                 1) // self.sampler.stride[0] + 1
+        out_w = (W + 2 * self.sampler.padding[1] - self.sampler.dilation[1] *
+                 (self.sampler.kernel_size[1] - 1) -
+                 1) // self.sampler.stride[1] + 1
+
+        output_size = (out_h, out_w)
         x = x.transpose(1, 2)  # B, H/2*W/2, 4*C
 
-        x = self.norm(x) if self.norm else x
-        x = self.reduction(x)
+        if self.is_post_norm:
+            # use post-norm here
+            x = self.reduction(x)
+            x = self.norm(x) if self.norm else x
+        else:
+            x = self.norm(x) if self.norm else x
+            x = self.reduction(x)
 
-        return x
+        return x, output_size
